nifty-social-signal/requirements.txt
tweepy==4.14.0
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3
nltk>=3.8.1
pyarrow>=14.0.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
